Use tags, and include more unclear terms
Expand dictionary to include terms likely to be referenced when predicting about AI safety approaches. 
(These will presumably not be used directly for forecasts, but may be used to forecasts about later human judgement. Example: In 2030, what is the probability that a panel of AI safety experts will unanimously agree that "Value Learning" is a solved problem?)